+++ asyncmy/replication/utils.py
@@ -5,8 +5,8 @@ def byte2int(b):
     if isinstance(b, int):
         return b
     else:
-        return B.unpack("!B", b)[0]
+        return B.unpack(b)[0]
 
 
 def int2byte(i: int):
-    return B.pack("!B", i)
+    return B.pack(i)