+++ asyncmy/protocol.pyx
@@ -117,7 +117,7 @@ cdef class MysqlPacket:
         self._position = end_pos + 1
         return result
 
-    cpdef unsigned long read_length_encoded_integer(self):
+    cpdef read_length_encoded_integer(self):
         """
         Read a 'Length Coded Binary' number from the data buffer.
 
@@ -126,7 +126,7 @@ cdef class MysqlPacket:
         """
         cdef int c = self.read_uint8()
         if c == NULL_COLUMN:
-            return 0
+            return None
         if c < UNSIGNED_CHAR_COLUMN:
             return c
         elif c == UNSIGNED_SHORT_COLUMN:
@@ -136,7 +136,7 @@ cdef class MysqlPacket:
         elif c == UNSIGNED_INT64_COLUMN:
             return self.read_uint64()
 
-    cpdef bytes read_length_coded_string(self):
+    cpdef read_length_coded_string(self):
         """
         Read a 'Length Coded String' from the data buffer.
 
@@ -145,8 +145,8 @@ cdef class MysqlPacket:
         that many bytes of binary data.  (For example "cat" would be "3cat".)
         """
         length = self.read_length_encoded_integer()
-        if length == 0:
-            return b""
+        if length is None:
+            return None
         return self.read(length)
 
     cpdef tuple read_struct(self, str fmt):