+++ asyncmy/replication/gtid.py
@@ -1,9 +1,8 @@
 import binascii
 import re
+import struct
 from io import BytesIO
 from typing import Set, Union
-
-import xstruct as struct
 
 
 class Gtid: