+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 0.2
 
+### 0.2.2
+
+- Fix bug. (#18)
+- Fix replication error.
+
 ### 0.2.1
 
 - Fix `binlogstream` await. (#12)